requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.28.0
typer>=0.9.0
python-dotenv>=1.0.0
groq>=0.4.0

# Development dependencies (optional, for code quality)
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
pytest>=7.4.0
types-python-dotenv>=1.0.0

